[TP1] Acomodadas las clases para resolver las diferentes estrategias del enunciado

+++ TP1/src/frsf/cidisi/exercise/tp1/search/actions/Avanzar.java
@@ -130,7 +130,7 @@ public class Avanzar extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(3000);
+        return new Double(1);
     }
 
     /**
