+++ TP1/src/frsf/cidisi/exercise/tp1/search/actions/GirarDer.java
@@ -78,7 +78,7 @@ public class GirarDer extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(1);
+        return new Double(30);
     }
 
     /**
